Release v0.50.0-alpha.0: Device config units display and grblHAL error codes
- Added unit field to firmware settings (mm/min, RPM, usec, etc.)
- Extended error decoder with grblHAL codes 39-75 and alarms 10-20
- Fixed non-numeric parameter editing (hostname, WiFi SSID, etc.)
- Updated all 130+ GRBL/grblHAL settings with unit labels

+++ crates/gcodekit5-devicedb/src/model.rs
@@ -112,9 +112,9 @@ pub struct DeviceProfile {
     pub timeout_ms: u64,
     pub auto_reconnect: bool,
 
-    /// Last known GRBL settings (from `$$`) for this profile.
+    /// Last known GRBL settings (from `$$`) for this profile (u16 to support grblHAL extended settings up to $680).
     #[serde(default)]
-    pub grbl_settings: std::collections::HashMap<u8, String>,
+    pub grbl_settings: std::collections::HashMap<u16, String>,
 }
 
 impl Default for DeviceProfile {
